requirements.txt
imutils==0.5.3
numpy==1.18.2
matplotlib==3.2.1
scikit-learn<=0.23.2
Pillow
streamlit
dtale